+++ ica-rs/src/data_struct/tailchat/status.rs
@@ -33,32 +33,18 @@ pub struct UpdateDMConverse {
     pub updated_at: String,
 }
 
+#[allow(unused)]
 pub type Writeable<T> = Arc<RwLock<T>>;
 
 #[derive(Debug, Clone)]
 pub struct BotStatus {
-    user_id: Writeable<UserId>,
+    user_id: UserId,
 }
 
 impl BotStatus {
-    pub fn new(user_id: UserId) -> Self {
-        Self {
-            user_id: Arc::new(RwLock::new(user_id)),
-        }
-    }
+    pub fn new(user_id: UserId) -> Self { Self { user_id } }
 
-    pub async fn get_user_id(&self) -> UserId { self.user_id.read().await.clone() }
-
-    pub fn block_get_user_id(&self) -> UserId {
-        tokio::task::block_in_place(|| {
-            let rt = tokio::runtime::Runtime::new().unwrap();
-            rt.block_on(self.get_user_id())
-        })
-    }
-
-    pub async fn set_user_id(&self, user_id: UserId) {
-        self.user_id.write().await.clone_from(&user_id);
-    }
+    pub fn get_user_id(&self) -> UserId { self.user_id.clone() }
 }
 
 #[derive(Debug, Clone)]