+++ ica-rs/src/py/call.rs
@@ -1,9 +1,11 @@
 use std::path::PathBuf;
+use std::sync::Arc;
 
 use pyo3::prelude::*;
 use rust_socketio::asynchronous::Client;
-use tracing::{info, warn};
+use tracing::{event, info, warn, Level};
 
+use crate::data_struct::tailchat::status::BotStatus;
 use crate::data_struct::{ica, tailchat};
 use crate::error::PyPluginError;
 use crate::py::{class, PyPlugin, PyStatus};
@@ -106,7 +108,7 @@ macro_rules! call_py_func {
                             $func_name.to_string(),
                             $plugin_path.to_string_lossy().to_string(),
                         );
-                        warn!("failed to call function<{}>: {:?}", $func_name, e);
+                        event!(Level::WARN, "failed to call function<{}>: {:?}", $func_name, e);
                     }
                 }
             })
@@ -126,20 +128,6 @@ pub async fn ica_new_message_py(message: &ica::messages::NewMessage, client: &Cl
         let args = (msg, client);
         // 甚至实际上压根不需要await这个spawn, 直接让他自己跑就好了(离谱)
         call_py_func!(args, plugin, path, ICA_NEW_MESSAGE_FUNC, client);
-        // tokio::spawn(async move {
-        //     Python::with_gil(|py| {
-        //         if let Ok(py_func) = get_func(plugin.py_module.bind(py), ICA_NEW_MESSAGE_FUNC) {
-        //             if let Err(e) = py_func.call1(args) {
-        //                 let e = PyPluginError::FuncCallError(
-        //                     e,
-        //                     ICA_NEW_MESSAGE_FUNC.to_string(),
-        //                     path.to_string_lossy().to_string(),
-        //                 );
-        //                 warn!("failed to call function<{}>: {:?}", ICA_NEW_MESSAGE_FUNC, e);
-        //             }
-        //         }
-        //     })
-        // });
     }
 }
 
@@ -152,29 +140,19 @@ pub async fn ica_delete_message_py(msg_id: ica::MessageId, client: &Client) {
         let client = class::ica::IcaClientPy::new(client);
         let args = (msg_id.clone(), client);
         call_py_func!(args, plugin, path, ICA_DELETE_MESSAGE_FUNC, client);
-        // tokio::spawn(async move {
-        //     Python::with_gil(|py| {
-        //         if let Ok(py_func) = get_func(plugin.py_module.bind(py), ICA_DELETE_MESSAGE_FUNC) {
-        //             if let Err(e) = py_func.call1(args) {
-        //                 let e = PyPluginError::FuncCallError(
-        //                     e,
-        //                     ICA_DELETE_MESSAGE_FUNC.to_string(),
-        //                     path.to_string_lossy().to_string(),
-        //                 );
-        //                 warn!("failed to call function<{}>: {:?}", ICA_DELETE_MESSAGE_FUNC, e);
-        //             }
-        //         }
-        //     })
-        // });
     }
 }
 
-pub async fn tailchat_new_message_py(message: &tailchat::messages::ReciveMessage, client: &Client) {
+pub async fn tailchat_new_message_py(
+    message: &tailchat::messages::ReceiveMessage,
+    client: &Client,
+    status: Arc<BotStatus>,
+) {
     verify_plugins();
 
     let plugins = PyStatus::get_files();
     for (path, plugin) in plugins.iter() {
-        let msg = class::tailchat::TailchatReciveMessagePy::from_recive_message(message);
+        let msg = class::tailchat::TailchatReceiveMessagePy::from_recive_message(message);
         let client = class::tailchat::TailchatClientPy::new(client);
         let args = (msg, client);
         call_py_func!(args, plugin, path, TAILCHAT_NEW_MESSAGE_FUNC, client);